pyproject.toml
[project]
name = "python-agentframework-demos"
version = "0.1.0"
description = "Collection of Python examples for Microsoft Agent Framework"
requires-python = ">=3.10"
dependencies = [
    "azure-identity",
    "openai>=1.109.1",
    "python-dotenv",
    "pydantic",
    "rich",
    "dotenv-azd",
    "aiohttp",
    "faker",
    "fastmcp",
    "agent-framework-core @ git+https://github.com/microsoft/agent-framework.git@65e77e52af5d99058be89823333f2f60afc07ab9#subdirectory=python/packages/core",
    "agent-framework-devui @ git+https://github.com/microsoft/agent-framework.git@65e77e52af5d99058be89823333f2f60afc07ab9#subdirectory=python/packages/devui",
    "agent-framework-redis @ git+https://github.com/microsoft/agent-framework.git@65e77e52af5d99058be89823333f2f60afc07ab9#subdirectory=python/packages/redis",
    "agent-framework-mem0 @ git+https://github.com/microsoft/agent-framework.git@65e77e52af5d99058be89823333f2f60afc07ab9#subdirectory=python/packages/mem0",
]

[tool.ruff]
line-length = 120
target-version = "py310"
lint.select = ["E", "F", "I", "UP"]
lint.ignore = ["D203"]
